Pulled set creation into utility file
